Refactor ozariaUtils
