+++ test/app/core/ozariaUtils.spec.js
@@ -1,4 +1,5 @@
-import ozariaUtils from '../../../app/core/ozariaUtils'
+/* eslint-env jasmine */
+import { findNextLevelsBySession } from 'ozaria/site/common/OzariaUtils'
 import factories from 'test/app/factories'
 import Levels from 'collections/Levels'
 
@@ -44,7 +45,7 @@ describe('ozaria utilities', () => {
       const levels = makeLevels(4)
       const sessions = makeLevelSessions(levels, [{ complete: true }, { complete: true }, { complete: false }])
       const expectedNextLevel = levels[2]
-      const nextLevelOriginal = ozariaUtils.findNextLevelsBySession(sessions, levels)
+      const nextLevelOriginal = findNextLevelsBySession(sessions, levels)
       expect(nextLevelOriginal.length).toBe(1)
       expect(nextLevelOriginal[0]).toEqual(expectedNextLevel.get('original'))
     })
@@ -53,7 +54,7 @@ describe('ozaria utilities', () => {
       const levels = makeLevels(4)
       const sessions = makeLevelSessions(levels, [{ complete: false }])
       const expectedNextLevel = levels[0]
-      const nextLevelOriginal = ozariaUtils.findNextLevelsBySession(sessions, levels)
+      const nextLevelOriginal = findNextLevelsBySession(sessions, levels)
       expect(nextLevelOriginal.length).toBe(1)
       expect(nextLevelOriginal[0]).toEqual(expectedNextLevel.get('original'))
     })
@@ -62,7 +63,7 @@ describe('ozaria utilities', () => {
       const levels = makeLevels(4)
       const sessions = makeLevelSessions(levels, [])
       const expectedNextLevel = levels.find((l) => l.get('first')) // first level will be next level
-      const nextLevelOriginal = ozariaUtils.findNextLevelsBySession(sessions, levels)
+      const nextLevelOriginal = findNextLevelsBySession(sessions, levels)
       expect(nextLevelOriginal.length).toBe(1)
       expect(nextLevelOriginal[0]).toEqual(expectedNextLevel.get('original'))
     })
